notification consumer

+++ faydh/lib/signin.dart
@@ -116,18 +116,18 @@ class _signInSreenState extends State<signInSreen> {
               FirebaseFirestore.instance
                   .collection('foodPost')
                   .doc(doc["docId"])
-                  .update({"expFlqag": FieldValue.increment(1)});
-            } else {
+                  .update({"expFlag": FieldValue.increment(1)});
+            } else if (doc["expFlag"] == 3) {
               FirebaseFirestore.instance
                   .collection('foodPost')
                   .doc(doc["docId"])
                   .delete();
+              FirebaseFirestore.instance
+                  .collection('users')
+                  .doc(doc["Cid"])
+                  .update({"ExpCount": FieldValue.increment(1)});
             }
           });
-          FirebaseFirestore.instance
-              .collection('users')
-              .doc(doc["Cid"])
-              .update({"ExpCount": FieldValue.increment(1)});
         } //end if
         if (doc["reserve"] == '1' && doc["notify"] == '0') {
           print('notify');
@@ -204,19 +204,18 @@ class _signInSreenState extends State<signInSreen> {
               FirebaseFirestore.instance
                   .collection('foodPost')
                   .doc(doc["docId"])
-                  .update({"expFlqag": FieldValue.increment(1)});
-            } else {
+                  .update({"expFlag": FieldValue.increment(1)});
+            } else if (doc["expFlag"] == 3) {
               FirebaseFirestore.instance
                   .collection('foodPost')
                   .doc(doc["docId"])
                   .delete();
+              FirebaseFirestore.instance
+                  .collection('users')
+                  .doc(doc["Cid"])
+                  .update({"ExpCount": FieldValue.increment(1)});
             }
           });
-
-          FirebaseFirestore.instance
-              .collection('users')
-              .doc(doc["Cid"])
-              .update({"ExpCount": FieldValue.increment(1)});
         }
       });
     });
